Add support for running WM with a docker Postgres container in addition to local postgres (#113)
This allows developers to run a docker postgres container for WM to connect to when running in local development mode. This is helpful so that the developer's local postgres environment doesn't need to match WM's preferred postgres setup. This is also helpful when a developer is working on multiple services locally (but not simultaneously), preventing any database name conflicts. Copied from Janitor, which was copied from Sam.

`./local_dev/run_postgres.sh start`

Co-authored-by: Isaac Zarsky <zarsky@broadinstitute.org>