[PF-199] Run integration test against localhost instead of dev (#127)
Currently, the integration test runs against the dev environment after the PR is merged, but before the new image is pushed to dev. This PR makes it so that the integration test for a PR is run against a local instance of the service with the PR changes.
Also change the integration test workflow to run on PRs as well as after merges to dev.