PF-186 Add spend authz check and billing account setup to create google context flight. (#147)
Check that a workspace has a valid spend profile id and that the user is authorized to link it before creating the google context. Set up that billing account on the project as a part of creating the google context. This allows the created project to spend money.
For now, only have tests with mocked sam spend link interaction.