PF-338 Add the service account user role temporarily for owners & writers (#219)
This role enables all owners and writers to "use" any service account in the project.
This is a temporary measure to enable AI Notebooks while we build APIs for controlled service accounts. Until then, to create AI Notebooks, a user needs to have the user role on a service account.