PF-469 Add AI Notebook Instance Controlled Resource Classes (#276)
Add a ControlledResource class for AI Notebook instances. Adds validation and serialization helpers for work in DAOs.
To be used in follow up PRs that actually create notebook instances.

Also move Bucket resource id creation from happening by default in the builder to explicitly in the Controller. This makes it harder to forget to set the resource id and makes the controller's creation logic happen in the controller.