This refactor is in preparation for Azure and in support of the PoC. I initially made these changes in the PoC branch, but they touch a lot of files that have nothing to do with the PoC.
There are three main changes. First, remove cloud context from workspace object. Having
it there makes the DAO complex. Also, most use cases in flights do not need to fetch the
workspace object, but can get the cloud context with the workspace id and use it to get
the GCP project id. Added helper methods in WorkspaceService for getting the context and
the project id.

Second, add a `creating_flight` column to the cloud_context table. Use that instead of
the project id for delete checking. Azure does not have a per-context unique id to check.
The flight id provides a general solution to ensure that the undo of a cloud context
create will only delete the context that it created.

Third, moved the serdes for GcpCloudContext into the GcpCloudContext class instead of
having it in the DAO. Seems tidier to me that way: cloud context knows how to serialize
and deserialize itself.