[WOR-483] Do not hardcode SAM resource type (#774)
Context

We should not hardcode the resource type when determining user permissions as we mint a SAS token. For controlled application private containers (such as those created by leo), this means the user will not be able to mint a token for said container.

Instead we should get it from the container resource instance.

This PR

Refactors the AzureControlledStorageResourceService in the following ways:
Renames to AzureStorageAccessService -- this feels like it's closer what the class actually does
Extracts the fetch of the storage account keys to a helper class StorageAccountKeyProvider. This code was responsible for pulling in several dependencies (AzureCloudContext, CrlService to name a few) that was making unit testing difficult
Creates a related StorageAccountKeyProviderUnitTest that isolates the more in-depth azure mocking
Pushes the endpoint string interpolation to the ControlledAzureStorageResource class
Converts the AzureControlledStorageAccessServiceTest integration test into a unit test (AzureStorageAccessServiceUnitTest) that removes the workspace and stairway dependencies and just exercises the functionality of the class