Add optional CORS support for azure storage accounts (#815)
Context

We need to set CORS allowed origins, methods and headers on Azure storage containers so calls from the browser succeed.

This PR

Wires up the call to Azure to set the needed values on the owning storage account at the time of creation. The allowed origins are config'ed out and default to none (these will be set via a subsequent terra-helmfile PR).