[WOR-530] Clone azure storage container definition (#883)
This is the first step in getting azure storage container cloning up and running. This commit is focused on the basics of wiring up a clone container flight that clones a container definition, deferring the actual copying of the constituent files to a subsequent ticket. There is also a follow-on task to create an integration test that exercises this flight.

This PR

* Creates a new CloneControlledAzureStorageContainerResourceFlight that copies a container's definition. Copying referenced containers is not supported, nor is the actual copying of files. I've stubbed out the step that will house the file copy operations for now.
* To determine the destination container's parent storage account, I've implemented a step called RetrieveDestinationStorageAccountResourceIdStep. This step determines the storage account to use for the container, either the destination workspace's own storage account (if present), or the destination workspace's parent landing zone (again, if present).
* Adds a test for the retrieval of the storage account, as well as the container definition creation. As noted, we'll implement an integration test once this flight is fully built.