[WOR-591] Build azure blob copying for cloning operations (#891)
Context

In this PR, we built out the Azure container cloning flight, stubbing out the copying of blobs for a separate effort. This PR builds out the blob copying step.

This PR

Builds out the CopyAzureStorageContainerBlobsStep. The step delegates mostly to a BlobCopier class that handles the mechanics of kicking off the blob copies and polling on their completion.
The BlobCopier starts the copy operations and polls on them to completion. For now, this is a serial operation. In the future we may consider a more sophisticated fork-join model similar to what Rawls does but I'd like to vet out this approach before building out more complex multithreaded code.
Adds a connected test that:
Creates two containers, uploads small test blobs to them, initiates a copy and verifies they were successfully copied.