[WOR-690] Remove storage account ID arg when creating destination storage container (#992)
We do not want to pass a storage account ID during cloning as we do not support non-landing-zone backed workspaces (for cloning). The storage account ID we are passing here is not a workspace manager resource ID but rather a landing zone ID that is tagged in azure. Omitting the storage account ID shifts to using the landing zone's storage account.

Related:
We should fail before dispatching the clone job if the destination does not have a landing zone. This is out of scope for this PR, but I've ticketed [here] for further consideration. (https://broadworkbench.atlassian.net/browse/WOR-695).