PF-1687 - Validate region constraints when adding data collections (#1049)
The operation for adding data collection resources into your workspace show as cloning referenced resources to your workspace.
When this happens, we need to check the policy on the data collection (workspace) and verify any region policies that exist there are compatible with region policies (if any) on the target workspace. Those need to be merged with the least privilege. Calculating that should be handled by the policy service.
When adding resources into the current workspace, we need to also verify that no resources already exist in the workspace that would violate the policy after merging.