[WOR-804] Handle VM extension provisioning errors (#1110)
Context

When a bad file URI is passed as an argument to the azure create VM step, azure will fail out with a VMExtensionProvisioningError.
This error comes back with an HTTP status code of 200. The generic ManagementException error handling logic only treats 4xx errors as FATAL.

So, WSM goes to retry the VM creation and then gets a 409 conflict as the moribund VM is still being destroyed. This has the undesirable effect of masking the root cause.

This PR

Adds explicit handling for the VMExtensionProvisioningError and immediately fails out the flight, which should make these issues a bit easier to diagnose in the future.
Re-enables the CreateAzureVmStepTest. We disabled this test in January for reasons that are murky at the moment (I think it may because we shifted to an explicit VM image name and the mocking was failing out). I've re-enabled the test and fixed up the mocking, as well as increased the coverage for the new error and another codepath that wasn't being exercised.