[WOR-993] Get user email from sam rather than relying on email from token (#1251)
We are relying on the email from the AuthenticatedUserRequest when logging the user email as we go through the process of minting a SAS token. This leads to inconsistent results as not all request tokens possess an email. We should use sam as the source of truth for the user's email.

Adds a call to Sam to fetch the user's email and threads it through into the relevant logging statements