[WOR-1006] Add content disposition parameter to SAS (#1303)
We would like to more easily tie the storage access logs in Azure together with the WSM logs as captured in GCP. Today that requires a SHA256 from the SAS request (or the corresponding blob log entry in Azure). This works, but is cumbersome. We are going to start including the Sam subject ID in the SAS so we can more easily search for logs related to a particular user in both systems.

This PR pulls the request user's subject ID from Sam and sets it as the [content disposition value](https://learn.microsoft.com/en-us/dotnet/api/azure.storage.sas.sasqueryparameters.contentdisposition?view=azure-dotnet) when minting a SAS token. This is technically not what the header is for, but given that we do not serve files directly in the browser, instead giving them SAS's and an azcopy cmd, this seems like an acceptable tradeoff.