[WOR-1430] Allow rawls-stage workspaces to be created with policies (#1601)
## Context

We want to be able to import snapshots to a RAWLS_STAGE workspace and have the policies attached to the snapshot merge with those of the workspace.

After testing out the code more as part of this PR, WSM already allows the snapshot to be imported and merges the policies as expected. But, there are still gates around creating Rawls workspaces with policies and the deletion of those policies at workspace deletion time. This PR has become an effort to "true up" the creation behavior with the snapshot import behavior and ensure that we don't have the artificial boundary at Rawls workspace creation time.

## This PR

* Removes the checks on the RAWLS_STAGE workspace at creation and deletion time
* Adds a test in the datarepo snapshot connected test to ensure that a group constraint policy propagates to the workspace
* There was some discussion about extending contract tests. Investigating further, I observe that:
There is no connection and thus no contract between TDR and WSM at the time of snapshot import. That is, WSM allows the addition of a reference with a snapshot ID but does not make a call out to TDR.