[WOR-1426] Shift to Repo secrets for some CI actions (#1619)
The write-config.sh script is liberally used throughout WSM CI. This means that GitHub needs to pull secrets from our Vault instance to run ITs, etc. We want to move away from this and leverage GH repo secrets instead (wired up via Terraform.

The original plan was to shift to pure workload identity federation via the google/auth github action. This proved unworkable for integration tests as there is a significant amount of machinery that relies on SA keys being placed in the /config directory with well-known names.

As initially mentioned, I've instead pursued a middle path of wiring the secrets into the repo, pulling them in the GitHub actions and rendering the keys at the needed well-known paths.