[WOR-1426] Fix full integration tests and service test workflows (#1625)
The nightly full integration test run includes ITs that are not typically included in PR integration tests. One that is consistently failing requires TPS to be enabled. WOR-1441 should mitigate that; until then I'm flipping the flag on by rendering a config file similar to what write-config.sh does (here is a successful run).

The nightly full service test flow requires Azure credentials to be in place (here is a link to failing run, note the failing Azure tests). This will require WOR-1463 to render the azure app credentials (or perhaps use a federated identity). I've switched this workflow back to using the old vault config until that ticket can be played. Here is a link to a run. One test on this run failed due to connectivity issues w/TPS, but note the passing Azure tests.