Refactor Stairway initialization out of JobService and into new StairwayInitializerService
This code was previously hard to unit test within the Spring Boot JobServiceTest.
We can now unit test that Stairway is initialized with a StairwayOptionsBuilder that is set up as we expect, including with all of our Stairway hooks (MdcHook and more).
We can also unit test StartupInitializer.initialize to make sure that it's performing all of the initialization we expect.