[WOR-1527] Remove Gradle code quality plugin
The current standard for static analysis in Workspaces repositories is to use Sonarcloud.
To match up with BPM and LZS, this commit removes the Gradle code quality plug-in from WSM.
Its configuration amounts to a no-op since we removed Spotbugs, and this clean-up makes for one fewer dependency for Dependabot to chase (currently, its attempts to update this plug-in are failing).