[WOR-1463] Remove vault dependency from staging promotion workflow (#1662)
The Release promotion tests workflow still has a vault dependency. This needs to be shifted to using github secrets. In my testing, I observed that this flow runs only in the staging environment so I've removed the parameterization as well. We should reflect on the need and contents of these promo tests going forward, but this is not that PR.