[WOR-1566] Upgrade io.kubernetes:client-java to 20.0.1
Breaking changes were introduced in v20.0.0: optional parameters are now consolidated into a single object, and we must explicitly execute our requests
(the client now supports different flavors of execution such as async).

This change was motivated by a desire to address a security vulnerability with jose4j, a transitive dependency of the Kubernetes client.
Upgrading the client brought this dependency past its vulnerability without needing to explicitly pin it.