[WOR-1589] Use the WSM SA token when pulling a landing zone's region. (#1690)
Context

Creation of Azure resources now requires a fetch of the landing zone region. This in turn requires permissions against the landing zone. This problematic as workspaces may be shared with users outside of a billing project who do not possess the requisite permissions to query the landing zone.

This PR

Shifts the calls to fetch the landing zone region to use the WSM SA token, as it should have administrative permissions against the landing zone.
Adjusts the checks in the azure resource creation controller to:
Validate the user's workspace and resource creation permissions first
Fetch the landing zone region for the needed common fields after the above checks