[WOR-1127] Wire up HTTP LZS client (#1710)
Context

This is the followup to #1709.
We want to be able to switch over to using the new Landing Zone web service when we are ready. The previous PR (linked above) introduced a new interface that abstracted over the direct library calls ("amalgamated"). The next step is to introduce an implementation of that interface for the LZS HTTP API and the needed configuration that will allows switching to that API when ready.

The LZS is disabled by default. 

This PR

* Introduces an HttpLandingZoneService implementation of the WorkspaceLandingZoneService interface. This is a "vanilla" WSM HTTP client layer, with the requisite LzsRetry class and internal exception hierarchy.
* Adds new type marshaling methods to the LandingZoneApiTypeAdapter to help map between Landing Zone Service API client types and the needed WSM API types for LandingZoneApiDispatch
* Adds a test for LandingZoneApiTypeAdapter
* I have not implemented a unit test for the HttpLandingZoneService class. My plan is to introduce coverage for that code via pact tests as described in WOR-1119.