[WOR-1671] Switch provider name (#1751)
Webhooks firing into LZS on can-i-deploy are failing. My current hypothesis is that this is related to the provider name being different in lzs and terra-helmfile ("terra-landing-zone-service" vs "landingzone"). I've deleted the pact in the broker so can-i-deploy should pass and updated the provider here. Next step is to update the provider name in LZS and publish a new pact.