[WOR-1507] Capture metrics and status for flights and steps #1753
We want to capture the success/failure rate of flights.

This PR introduces a FlightMetricsHook class. This is a stairway lifecycle hook that emits a micrometer timer at the end of steps and flights. The metrics are namespaced in the following manner:

For flight-level data:

stairway_flight_<flight_class_name>_duration_seconds_<metric>
For step-level data:

stairway_flight_<flight_class_name>_step_<step_class_name>_duration_seconds_<metric>
The following tags are attached to both types of metrics:

Status: Final status of the flight
Resource type: Cloud resource type; this is only emitted in the case of flights with a RESOURCE_TYPE value in their input parameters (i.e,. CreateControlledCloudResourceFlight). It is omitted otherwise.