[WOR-1799] Verify policy service and buffer service credentials and register WSM SA at bootstrap (#1780)
In testing Helmfile changes to credentials configuration, it was difficult to tell if WSM external service credentials were configured properly without firing live requests at it (i.e,. the service configuration may be valid yaml, but I may have introduced a typo in the credentials path, which did not manifest until actual live workspace traffic was fired at the application resulting in a 500)

This PR

Verifies that credentials are in the right place during application bootstrap.
Register the WSM SA with Sam (if needed) during bootstrap, rather than waiting for the creation of a controlled resource.
I considered adding TPS + buffer to the WSM status endpoint checks; I've decided not to as I view those as important but non-critical services for a k8s status check. I've left the status hooks in the relevant service classes so those can be easily wired up in the future.